refactor(ai): extract orchestrator workspace staging
