+++ src/internal/ai/orchestrator/executor.rs
@@ -1,6 +1,6 @@
 use std::{
-    collections::{BTreeMap, BTreeSet, HashMap, HashSet},
-    fs, io,
+    collections::{HashMap, HashSet},
+    io,
     path::{Path, PathBuf},
     sync::Arc,
     time::Duration,
@@ -11,7 +11,6 @@ use dagrs::{
     Action, CheckpointConfig, DefaultNode, EnvVar, FileCheckpointStore, Graph, InChannels, Node,
     NodeTable, OutChannels, Output, event::GraphEvent,
 };
-use git_internal::hash::ObjectHash;
 use serde::Deserialize;
 use serde_json::Value;
 use tokio::sync::Semaphore;
@@ -26,9 +25,9 @@ use super::{
         ExecutionPlanSpec, GateReport, OrchestratorError, OrchestratorObserver, ReviewOutcome,
         TaskKind, TaskNodeStatus, TaskResult, TaskSpec, ToolCallRecord,
     },
+    workspace::{cleanup_task_worktree, prepare_task_worktree, sync_task_worktree_back},
 };
 use crate::{
-    command::calc_file_blob_hash,
     internal::ai::{
         agent::{ToolLoopConfig, ToolLoopObserver, run_tool_loop_with_history_and_observer},
         completion::{CompletionError, CompletionModel},
@@ -157,16 +156,6 @@ struct ReviewerDecision {
     summary: String,
     #[serde(default)]
     issues: Vec<String>,
-}
-
-#[derive(Clone, Debug, Default)]
-struct WorkspaceSnapshot {
-    files: BTreeMap<PathBuf, ObjectHash>,
-}
-
-struct TaskWorktree {
-    root: PathBuf,
-    baseline: WorkspaceSnapshot,
 }
 
 /// Execute a single task with retry logic.
@@ -564,182 +553,6 @@ fn task_workspace_failure(task: &TaskSpec, err: io::Error) -> TaskResult {
         policy_violations: Vec::new(),
         review: None,
     }
-}
-
-fn prepare_task_worktree(main_working_dir: &Path, task_id: Uuid) -> io::Result<TaskWorktree> {
-    let storage = util::try_get_storage_path(Some(main_working_dir.to_path_buf()))?;
-    let root = std::env::temp_dir().join(format!(
-        "libra-task-worktree-{}-{}",
-        std::process::id(),
-        task_id
-    ));
-
-    if root.exists() {
-        fs::remove_dir_all(&root)?;
-    }
-    fs::create_dir_all(&root)?;
-    create_storage_link(&storage, &root.join(util::ROOT_DIR))?;
-
-    let baseline = snapshot_workspace(main_working_dir)?;
-    materialize_workspace(main_working_dir, &root, &baseline)?;
-
-    Ok(TaskWorktree { root, baseline })
-}
-
-fn cleanup_task_worktree(root: &Path) -> io::Result<()> {
-    if root.exists() {
-        fs::remove_dir_all(root)?;
-    }
-    Ok(())
-}
-
-fn snapshot_workspace(root: &Path) -> io::Result<WorkspaceSnapshot> {
-    fn visit_dir(
-        root: &Path,
-        dir: &Path,
-        files: &mut BTreeMap<PathBuf, ObjectHash>,
-    ) -> io::Result<()> {
-        for entry in fs::read_dir(dir)? {
-            let entry = entry?;
-            if entry.file_name() == util::ROOT_DIR {
-                continue;
-            }
-
-            let path = entry.path();
-            let metadata = fs::metadata(&path)?;
-            if metadata.is_dir() {
-                visit_dir(root, &path, files)?;
-                continue;
-            }
-
-            let rel = path
-                .strip_prefix(root)
-                .map_err(|err| io::Error::other(err.to_string()))?
-                .to_path_buf();
-            files.insert(rel, calc_file_blob_hash(&path)?);
-        }
-        Ok(())
-    }
-
-    let mut files = BTreeMap::new();
-    visit_dir(root, root, &mut files)?;
-    Ok(WorkspaceSnapshot { files })
-}
-
-fn materialize_workspace(
-    source_root: &Path,
-    target_root: &Path,
-    snapshot: &WorkspaceSnapshot,
-) -> io::Result<()> {
-    for rel_path in snapshot.files.keys() {
-        copy_workspace_file(source_root, target_root, rel_path)?;
-    }
-    Ok(())
-}
-
-fn sync_task_worktree_back(
-    main_working_dir: &Path,
-    task_worktree_dir: &Path,
-    baseline: &WorkspaceSnapshot,
-) -> io::Result<()> {
-    let task_snapshot = snapshot_workspace(task_worktree_dir)?;
-    let changed_paths = changed_paths_since_baseline(baseline, &task_snapshot);
-
-    for rel_path in &changed_paths {
-        let expected = baseline.files.get(rel_path).copied();
-        let actual = file_hash_if_exists(&main_working_dir.join(rel_path))?;
-        if actual != expected {
-            return Err(io::Error::other(format!(
-                "main workspace changed concurrently at '{}'",
-                rel_path.display()
-            )));
-        }
-    }
-
-    for rel_path in changed_paths {
-        if task_snapshot.files.contains_key(&rel_path) {
-            copy_workspace_file(task_worktree_dir, main_working_dir, &rel_path)?;
-        } else {
-            remove_workspace_file(main_working_dir, &rel_path)?;
-        }
-    }
-
-    Ok(())
-}
-
-fn changed_paths_since_baseline(
-    baseline: &WorkspaceSnapshot,
-    current: &WorkspaceSnapshot,
-) -> Vec<PathBuf> {
-    let paths = baseline
-        .files
-        .keys()
-        .chain(current.files.keys())
-        .cloned()
-        .collect::<BTreeSet<_>>();
-
-    paths
-        .into_iter()
-        .filter(|path| baseline.files.get(path) != current.files.get(path))
-        .collect()
-}
-
-fn file_hash_if_exists(path: &Path) -> io::Result<Option<ObjectHash>> {
-    if !path.exists() {
-        return Ok(None);
-    }
-    Ok(Some(calc_file_blob_hash(path)?))
-}
-
-fn copy_workspace_file(source_root: &Path, target_root: &Path, rel_path: &Path) -> io::Result<()> {
-    let source = source_root.join(rel_path);
-    let target = target_root.join(rel_path);
-    if let Some(parent) = target.parent() {
-        fs::create_dir_all(parent)?;
-    }
-    fs::copy(&source, &target)?;
-    let permissions = fs::metadata(&source)?.permissions();
-    fs::set_permissions(&target, permissions)?;
-    Ok(())
-}
-
-fn remove_workspace_file(root: &Path, rel_path: &Path) -> io::Result<()> {
-    let target = root.join(rel_path);
-    if target.exists() {
-        fs::remove_file(&target)?;
-        remove_empty_parents(root, target.parent());
-    }
-    Ok(())
-}
-
-fn remove_empty_parents(root: &Path, mut current: Option<&Path>) {
-    while let Some(dir) = current {
-        if dir == root {
-            break;
-        }
-
-        let is_empty = match fs::read_dir(dir) {
-            Ok(mut entries) => entries.next().is_none(),
-            Err(_) => false,
-        };
-        if !is_empty {
-            break;
-        }
-        if fs::remove_dir(dir).is_err() {
-            break;
-        }
-        current = dir.parent();
-    }
-}
-
-#[cfg(unix)]
-fn create_storage_link(storage: &Path, link_path: &Path) -> io::Result<()> {
-    std::os::unix::fs::symlink(storage, link_path)
-}
-
-#[cfg(windows)]
-fn create_storage_link(storage: &Path, link_path: &Path) -> io::Result<()> {
-    std::os::windows::fs::symlink_dir(storage, link_path)
 }
 
 struct TaskDagrsAction<M: CompletionModel + 'static> {
